Update README.md
The code I provided is a basic implementation of the Facebook Login feature using the Facebook SDK for PHP. Here's a breakdown of the code and its functionality:

config.php:

·	Defines the Facebook App ID and Secret as constants
·	Loads the Facebook SDK autoload file

login.php:

·	Includes the config.php file to access the App ID and Secret
·	Creates a new Facebook instance with the App ID and Secret
·	Gets an instance of the RedirectLoginHelper class
·	Generates a login URL with the required permissions (email and public profile)
·	Outputs an HTML link to the login URL

When a user clicks the link, they will be redirected to the Facebook login page. After authorization, Facebook will redirect the user back to your website. The RedirectLoginHelper class handles the login flow and retrieves an access token for the user.

This is a basic implementation, and you may need to add additional functionality, such as handling the login response, retrieving user data, and managing access tokens.


Signed-off-by: mrlarge <152041668+mrlarge0001@users.noreply.github.com>